+++ TDD/mycode.py
@@ -1,25 +1,20 @@
-import unittest
-import mycode
+def add(x, y):
+    return x + y
 
-class Tests(unittest.TestCase):
+def multiply(x, y):
+    return x * y
 
-    def test_add(self):
-        self.assertEqual(mycode.add(6,6),12)
-        self.assertEqual(mycode.add(-1,2),1)
-        self.assertEqual(mycode.add(-5,2),-3)
+def subtract(x, y):
+    return x - y
+    # if "-" in str(x):
+    #     x = 0 - x
+    # elif "-" in str(y):
+    #     y = 0 - x
+    # return x - y
 
-    def test_multiply(self):
-        self.assertEqual(mycode.multiply(6,3),18)
-        self.assertEqual(mycode.multiply(1,1),1)
-        self.assertEqual(mycode.multiply(5,2),10)
-
-    def test_subtract(self):
-        self.assertEqual(mycode.add(6,6),0)
-        self.assertEqual(mycode.add(-1,-1),2)
-        self.assertEqual(mycode.add(5,2),3)
-
-    def test_divide(self):
-        self.assertEqual(mycode.divide(6,6),1)
-        self.assertEqual(mycode.divide(12,2),6)
-        self.assertEqual(mycode.divide(5,2),2.5)
-        self.assertEqual(mycode.divide(5/0),"Invalid")
+def divide(x, y):
+    if x == 0 or y == 0:
+        diff = "Invalid"
+    else:
+        diff = x / y
+    return diff
